refactor(app): добавил комментарии, линтер, зависимости, разложил по папкам.
