+++ docs/RESULT.md
@@ -5,6 +5,8 @@
 # Результаты
 
 [здесь можете похвастаться, что успели реализовать, или сделали что-то сверх задания]
+Дополнительно:
+- Слежу за чистотой коммитов, чтобы по ним можно было отследить CHANGELOG.
 
 # Ссылки на демонстрацию работы/скриншоты
