+++ lib/core/my_app.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:surf_flutter_study_jam_2023/features/ticket_storage/ticket_storage_page.dart';
+
+/// Корневой виджет приложения.
+class MyApp extends StatelessWidget {
+  /// @nodoc
+  const MyApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Async Tour',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const TicketStoragePage(),
+    );
+  }
+}
