+++ lib/features/ticket_storage/ticket_storage_page.dart
@@ -2,7 +2,8 @@ import 'package:flutter/material.dart';
 
 /// Экран “Хранения билетов”.
 class TicketStoragePage extends StatelessWidget {
-  const TicketStoragePage({Key? key}) : super(key: key);
+  /// @nodoc
+  const TicketStoragePage({super.key});
 
   @override
   Widget build(BuildContext context) {
