+++ lib/main.dart
@@ -1,21 +1,11 @@
-import 'package:flutter/material.dart';
-import 'package:surf_flutter_study_jam_2023/features/ticket_storage/ticket_storage_page.dart';
+// TODO(zemcov): установи FVM.
 
+import 'package:surf_flutter_study_jam_2023/runner.dart';
+
+/// Точка входа в приложение.
+///
+/// Так как точек входа может быть много, в [main] удобно
+/// инициализировать переменные окружения.
 void main() {
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: const TicketStoragePage(),
-    );
-  }
+  runner();
 }
